Split helper.js into render.js and helper.js
