+++ src/core/index.js
@@ -1,4 +1,4 @@
-const helper = require('./helper.js');
+const popperRenderer = require('./render.js');
 
 //Create a popper object (This is for use on the core)
 module.exports = function(userOptions){
@@ -7,12 +7,12 @@ module.exports = function(userOptions){
   var container = cy.container();
 
   //Generate options and assign them on the scratchpad 
-  var options = helper.createPopperOptionsObject(userOptions);
+  var options = popperRenderer.createPopperOptionsObject(userOptions);
   cy.scratch('popper-opts', options.popper);
   cy.scratch('popper-target', options.target);
 
   //Create popper object
-  var popper  = helper.createPopperObject(cy);
+  var popper  = popperRenderer.createPopperObject(cy);
   cy.scratch('popper', popper);
 
   return this;
